+++ src/packages/piral/src/components/Modals.tsx
@@ -2,6 +2,10 @@ import * as React from 'react';
 import { useGlobalState, OpenModalDialog } from 'piral-core';
 import { ModalsContainerProps } from '../types';
 
+function closeAll(modals: Array<OpenModalDialog>) {
+  modals.forEach(m => m.close());
+}
+
 export interface ModalsCreator {
   ModalsContainer: React.ComponentType<ModalsContainerProps>;
   ModalDialog: React.ComponentType<OpenModalDialog>;
@@ -10,9 +14,11 @@ export interface ModalsCreator {
 export function createModals({ ModalsContainer, ModalDialog }: ModalsCreator): React.SFC {
   return () => {
     const modals = useGlobalState(s => s.app.modals);
+    const open = modals.length > 0;
+    const close = () => closeAll(modals);
 
     return (
-      <ModalsContainer>
+      <ModalsContainer open={open} close={close}>
         {modals.map(n => (
           <ModalDialog {...n} key={n.name} />
         ))}